backend/requirements.txt
dedalus-labs
python-dotenv
mcp
